requirements.txt
alembic==1.11.1
apispec==5.2.2
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
dnspython==2.3.0
email-validator==1.3.1
Flask==2.3.2
Flask-AppBuilder==4.3.2
Flask-Babel==2.0.0
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
limits==3.5.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.26.1
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
pkgutil_resolve_name==1.3.10
prison==0.2.1
psycopg2-binary==2.9.6
psycopg2-pool==1.1
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
rich==13.3.5
six==1.16.0
SPARQLWrapper==2.0.0
SQLAlchemy==1.4.48
SQLAlchemy-Utils==0.41.1
typing_extensions==4.6.0
urllib3==2.0.2
Werkzeug==2.3.4
wrapt==1.15.0
WTForms==3.0.1
zipp==3.15.0
